Adjust logic to return 404 on bad redeem key, attempt to fix test instability
