+++ api/src/controllers/events.rs
@@ -677,25 +677,23 @@ pub fn redeem_ticket(
     let db_event = Event::find(parameters.id, connection)?;
     let organization = db_event.organization(connection)?;
     auth_user.requires_scope_for_organization_event(Scopes::RedeemTicket, &organization, &db_event, connection)?;
-    match TicketInstance::find_by_event_id_redeem_key(parameters.id, redeem_parameters.redeem_key.clone(), connection)
-        .optional()?
-    {
-        Some(ticket) => {
-            let redeemable = TicketInstance::show_redeemable_ticket(ticket.id, connection)?;
+    let ticket =
+        TicketInstance::find_by_event_id_redeem_key(parameters.id, redeem_parameters.redeem_key.clone(), connection)?;
+    let redeemable = TicketInstance::show_redeemable_ticket(ticket.id, connection)?;
 
-            let result = TicketInstance::redeem_ticket(
-                ticket.id,
-                redeem_parameters.redeem_key.clone(),
-                auth_user.id(),
-                redeem_parameters.check_in_source.unwrap_or(CheckInSource::GuestList),
-                connection,
-            )?;
+    let result = TicketInstance::redeem_ticket(
+        ticket.id,
+        redeem_parameters.redeem_key.clone(),
+        auth_user.id(),
+        redeem_parameters.check_in_source.unwrap_or(CheckInSource::GuestList),
+        connection,
+    )?;
 
-            match result {
-                RedeemResults::TicketRedeemSuccess => {
-                    //Redeem ticket on chain
-                    let asset = Asset::find(ticket.asset_id, connection)?;
-                    match asset.blockchain_asset_id {
+    match result {
+        RedeemResults::TicketRedeemSuccess => {
+            //Redeem ticket on chain
+            let asset = Asset::find(ticket.asset_id, connection)?;
+            match asset.blockchain_asset_id {
                         Some(a) => {
                             let wallet = Wallet::find(ticket.wallet_id, connection)?;
                             state.config.tari_client.modify_asset_redeem_token(&wallet.secret_key, &wallet.public_key,
@@ -720,20 +718,19 @@ pub fn redeem_ticket(
                         }
                         None => Ok(HttpResponse::BadRequest().json(json!({ "error": "Could not complete this checkout because the asset has not been assigned on the blockchain.".to_string()}))),
                     }
-                }
-                RedeemResults::TicketTransferInProcess => Ok(HttpResponse::BadRequest()
-                    .json(json!({"error": "Ticket has pending transfer in progress.".to_string()}))),
-                RedeemResults::TicketAlreadyRedeemed => Ok(HttpResponse::Conflict().json(json!({
-                "error": "Ticket has already been redeemed.".to_string(),
-                "redeemed_by": redeemable.redeemed_by,
-                "redeemed_at": redeemable.redeemed_at
-                }))),
-                RedeemResults::TicketInvalid => {
-                    Ok(HttpResponse::BadRequest().json(json!({"error": "Ticket is invalid.".to_string()})))
-                }
-            }
         }
-        None => return Ok(HttpResponse::BadRequest().json(json!({"error": "Ticket is invalid.".to_string()}))),
+        RedeemResults::TicketTransferInProcess => {
+            Ok(HttpResponse::BadRequest()
+                .json(json!({"error": "Ticket has pending transfer in progress.".to_string()})))
+        }
+        RedeemResults::TicketAlreadyRedeemed => Ok(HttpResponse::Conflict().json(json!({
+        "error": "Ticket has already been redeemed.".to_string(),
+        "redeemed_by": redeemable.redeemed_by,
+        "redeemed_at": redeemable.redeemed_at
+        }))),
+        RedeemResults::TicketInvalid => {
+            Ok(HttpResponse::BadRequest().json(json!({"error": "Ticket is invalid.".to_string()})))
+        }
     }
 }
