+++ api/tests/functional/base/events.rs
@@ -1,6 +1,7 @@
 use actix_web::{http::StatusCode, FromRequest, HttpResponse, Path, Query};
 use bigneon_api::controllers::events;
 use bigneon_api::controllers::events::*;
+use bigneon_api::db::CacheDatabase;
 use bigneon_api::extractors::*;
 use bigneon_api::models::*;
 use bigneon_db::dev::times;
@@ -62,6 +63,69 @@ pub fn clone(role: Roles, should_test_succeed: bool) {
             cloned_event.event_end.unwrap().timestamp(),
             clone_fields.event_end.timestamp()
         );
+    } else {
+        support::expects_unauthorized(&response);
+    }
+}
+
+pub fn redeem_ticket(role: Roles, should_test_succeed: bool) {
+    let database = TestDatabase::new();
+    let conn = database.connection.get();
+    let user = database.create_user().finish();
+    let request = TestRequest::create_with_uri_custom_params("/", vec!["id", "ticket_instance_id"]);
+    let organization = database.create_organization().finish();
+    let event = database
+        .create_event()
+        .with_organization(&organization)
+        .with_ticket_pricing()
+        .finish();
+    let user2 = database.create_user().finish();
+    let ticket_type = event.ticket_types(true, None, conn).unwrap()[0].id;
+    let ticket = database.create_purchased_tickets(&user2, ticket_type, 5).remove(0);
+    let auth_user = support::create_auth_user_from_user(&user, role, Some(&organization), &database);
+
+    let mut path = Path::<PathParameters>::extract(&request.request).unwrap();
+    path.id = event.id;
+    let mut path2 = Path::<PathParameters>::extract(&request.request).unwrap();
+    path2.id = event.id;
+
+    //First try when Redeem code is wrong
+    let request_data = TicketRedeemRequest {
+        redeem_key: "WrongKey".to_string(),
+        check_in_source: Some(CheckInSource::Scanned),
+    };
+
+    let response: HttpResponse = events::redeem_ticket((
+        database.connection.clone().into(),
+        path,
+        Json(request_data),
+        auth_user.clone(),
+        request.extract_state(),
+        CacheDatabase { inner: None },
+    ))
+    .into();
+
+    if should_test_succeed {
+        assert_eq!(response.status(), StatusCode::NOT_FOUND);
+        //Now try with redeem code being correct
+        let request_data = TicketRedeemRequest {
+            redeem_key: ticket.redeem_key.unwrap(),
+            check_in_source: Some(CheckInSource::Scanned),
+        };
+
+        let response: HttpResponse = events::redeem_ticket((
+            database.connection.clone().into(),
+            path2,
+            Json(request_data),
+            auth_user,
+            request.extract_state(),
+            CacheDatabase { inner: None },
+        ))
+        .into();
+        let ticket = TicketInstance::find(ticket.id, conn).unwrap();
+        assert_eq!(ticket.check_in_source, Some(CheckInSource::Scanned));
+
+        assert_eq!(response.status(), StatusCode::OK);
     } else {
         support::expects_unauthorized(&response);
     }
